nvidia-orin: pass version parameter for libnvidia-container
Issue: LINUXEXEC-35040

When building package libnvidia-container, there is an error reported.
Makefile:106: *** Invalid major version.  Stop.
Because during building this package, it needs a LIB_VERSION variable.
If it is not set explicitly, makefile set it by getting the version from
tag. But in WRLinux LTS25 product, gitshallow tarball is used, there is
no git info (for example, tags) in source code directory, so error is
reported. In order to fix this issue, add a bbappend to set version
explicitly.

Signed-off-by: Meng Li <Meng.Li@windriver.com>